+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/advancedsettings/AdvancedSettingsState.kt
@@ -1,5 +1,7 @@
 package co.electriccoin.zcash.ui.screen.advancedsettings
 
+import co.electriccoin.zcash.ui.design.component.ButtonState
+
 data class AdvancedSettingsState(
     val onBack: () -> Unit,
     val onRecoveryPhraseClick: () -> Unit,
@@ -7,4 +9,5 @@ data class AdvancedSettingsState(
     val onChooseServerClick: () -> Unit,
     val onCurrencyConversionClick: () -> Unit,
     val onDeleteZashiClick: () -> Unit,
+    val coinbaseButton: ButtonState?,
 )