+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/advancedsettings/AndroidAdvancedSettings.kt
@@ -2,11 +2,14 @@
 
 package co.electriccoin.zcash.ui.screen.advancedsettings
 
+import android.net.Uri
 import androidx.activity.compose.BackHandler
+import androidx.browser.customtabs.CustomTabsIntent
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.di.koinActivityViewModel
+import co.electriccoin.zcash.ui.common.compose.LocalActivity
 import co.electriccoin.zcash.ui.common.compose.LocalNavController
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.screen.advancedsettings.view.AdvancedSettings
@@ -20,6 +23,7 @@ internal fun WrapAdvancedSettings(
     goExportPrivateData: () -> Unit,
     goSeedRecovery: () -> Unit,
 ) {
+    val activity = LocalActivity.current
     val navController = LocalNavController.current
     val walletViewModel = koinActivityViewModel<WalletViewModel>()
     val viewModel = koinViewModel<AdvancedSettingsViewModel>()
@@ -38,6 +42,18 @@ internal fun WrapAdvancedSettings(
     LaunchedEffect(Unit) {
         viewModel.navigationCommand.collect {
             navController.navigate(it)
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        viewModel.coinbaseNavigationCommand.collect { uri ->
+            val intent =
+                CustomTabsIntent.Builder()
+                    .setUrlBarHidingEnabled(true)
+                    .setShowTitle(true)
+                    .setShareState(CustomTabsIntent.SHARE_STATE_OFF)
+                    .build()
+            intent.launchUrl(activity, Uri.parse(uri))
         }
     }