+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 - Transaction resubmission feature has been added. It periodically searches for unmined sent transactions that 
   are still within their expiry window and resubmits them if there are any.
 - The Choose server screen now provides a new search for the three fastest servers feature
+- Coinbase integration button added to Advanced Settings
 
 ### Changed
 - Choose server screen has been redesigned